requirements.txt
Flask==3.1.2
PyDrive2==1.21.3
python-dotenv==1.0.0
oauth2client==4.1.3
google-api-python-client==2.184.0
google-auth-httplib2==0.2.0
google-auth==2.41.1
Werkzeug>=3.1.0
gunicorn